requirements.txt
pandas
numpy
scikit-learn
plotly
prophet
cmdstanpy==1.2.0
holidays==0.29
tqdm==4.66.1
firebase-admin
psycopg2
python-dotenv